# Conflicts:
#	OdoriRails/OdoriRails.sln
#	OdoriRails/OdoriRails/DAL/MSSQLDatabaseContext.cs
#	OdoriRails/OdoriRails/DAL/MySqlContext.cs

+++ OdoriRails/OdoriRails/DAL/MSSQLDatabaseContext.cs
@@ -44,7 +44,8 @@ namespace OdoriRails.DAL
 
         public void RemoveUser(User user)
         {
-            GetData(new SqlCommand("DELETE FROM [User] WHERE UserPk = " + user.Id));
+            GetData(new SqlCommand($"DELETE FROM [User] WHERE UserPk = {user.Id}"));
+            GetData(new SqlCommand($"UPDATE [User] SET ManagedByFk = null WHERE ManagedByFk = {user.Id}"));
         }
 
         public User GetUser(int id)
@@ -202,7 +203,7 @@ FROM Clean INNER JOIN
 FROM Service INNER JOIN
 (SELECT ServiceUser.ServiceCk
 FROM ServiceUser INNER JOIN
-[User] ON ServiceUser.UserCk = [User].UserPk
+[User] ON ServiceUser.UserCk = [User].UserP
 WHERE ([User].UserPk = @id)) AS derivedtbl_1 ON Service.ServicePk = derivedtbl_1.ServiceCk) AS derivedtbl_2 ON Clean.ServiceFk = derivedtbl_2.ServicePk");
             cleanQuery.Parameters.AddWithValue("@id", user.Id);
             return GenerateListWithFunction(GetData(cleanQuery), CreateCleaning);
@@ -232,20 +233,21 @@ WHERE (ServiceUser.UserCk IS NULL)) AS derivedtbl_1 ON Clean.ServiceFk = derived
         {
             var serviceQuery = new SqlCommand(@"INSERT INTO Service (StartDate, EndDate, TramFk) VALUES (@startdate, @enddate, @tramfk); SELECT SCOPE_IDENTITY();");
             serviceQuery.Parameters.AddWithValue("@startdate", cleaning.StartDate);
-            if (cleaning.StartDate == DateTime.MinValue) serviceQuery.Parameters.AddWithValue("@enddate", DBNull.Value);
+            if (cleaning.EndDate == DateTime.MinValue) serviceQuery.Parameters.AddWithValue("@enddate", DBNull.Value);
             else serviceQuery.Parameters.AddWithValue("@enddate", cleaning.EndDate);
             serviceQuery.Parameters.AddWithValue("@tramfk", cleaning.TramId);
 
             var data = GetData(serviceQuery);
 
-            var cleaningQuery = new SqlCommand(@"INSERT INTO Cleaning (ServiceFk, Size, Remarks) VALUES (@id, @size, @remarks)");
-            cleaningQuery.Parameters.AddWithValue("@id", (int)data.Rows[0].ItemArray[0]);
+            var cleaningQuery = new SqlCommand(@"INSERT INTO Clean (ServiceFk, Size, Remarks) VALUES (@id, @size, @remarks)");
+            cleaningQuery.Parameters.AddWithValue("@id", data.Rows[0].ItemArray[0]);
             cleaningQuery.Parameters.AddWithValue("@size", (int)cleaning.Size);
-            cleaningQuery.Parameters.AddWithValue("@remarks", cleaning.Comments);
+            cleaningQuery.Parameters.AddWithValue("@remarks", cleaning.Comments ?? "");
+            GetData(cleaningQuery);
 
-            cleaning.SetId((int)data.Rows[0].ItemArray[0]);
             SetUsersToServices(cleaning.AssignedUsers, cleaning);
-            
+
+            cleaning.SetId(Convert.ToInt32((decimal)data.Rows[0].ItemArray[0]));
             return cleaning;
         }
 
@@ -253,22 +255,22 @@ WHERE (ServiceUser.UserCk IS NULL)) AS derivedtbl_1 ON Clean.ServiceFk = derived
         {
             var serviceQuery = new SqlCommand(@"INSERT INTO Service (StartDate, EndDate, TramFk) VALUES (@startdate, @enddate, @tramfk); SELECT SCOPE_IDENTITY();");
             serviceQuery.Parameters.AddWithValue("@startdate", repair.StartDate);
-            serviceQuery.Parameters.AddWithValue("@enddate", repair.EndDate);
+            if (repair.EndDate == DateTime.MinValue) serviceQuery.Parameters.AddWithValue("@enddate", DBNull.Value);
+            else serviceQuery.Parameters.AddWithValue("@enddate", repair.EndDate);
             serviceQuery.Parameters.AddWithValue("@tramfk", repair.TramId);
 
             var data = GetData(serviceQuery);
 
             var repairQuery = new SqlCommand(@"INSERT INTO Repair (ServiceFk, Solution, Defect, Type) VALUES (@id, @solution, @defect, @type)");
             repairQuery.Parameters.AddWithValue("@id", data.Rows[0].ItemArray[0]);
-            repairQuery.Parameters.AddWithValue("@solution", repair.Solution);
-            repairQuery.Parameters.AddWithValue("@defect", repair.Defect);
+            repairQuery.Parameters.AddWithValue("@solution", repair.Solution ?? "");
+            repairQuery.Parameters.AddWithValue("@defect", repair.Defect ?? "");
             repairQuery.Parameters.AddWithValue("@type", (int)repair.Type);
             GetData(repairQuery);
 
-
-            repair.SetId(Convert.ToInt32(data.Rows[0].ItemArray[0]));
             SetUsersToServices(repair.AssignedUsers, repair);
-            
+
+            repair.SetId(Convert.ToInt32((decimal)data.Rows[0].ItemArray[0]));
             return repair;
         }
 
@@ -288,8 +290,8 @@ WHERE (ServiceUser.UserCk IS NULL)) AS derivedtbl_1 ON Clean.ServiceFk = derived
                     {
                         var repair = (Repair)service;
                         var repairQuery = new SqlCommand("UPDATE Repair SET Solution = @solution, Defect = @defect, Type = @type WHERE RepairFK = @id");
-                        repairQuery.Parameters.AddWithValue("@solution", repair.Solution);
-                        repairQuery.Parameters.AddWithValue("@defect", repair.Defect);
+                        repairQuery.Parameters.AddWithValue("@solution", repair.Solution ?? "");
+                        repairQuery.Parameters.AddWithValue("@defect", repair.Defect ?? "");
                         repairQuery.Parameters.AddWithValue("@type", (int)repair.Type);
                         repairQuery.Parameters.AddWithValue("@id", repair.Id);
                         GetData(repairQuery);
@@ -300,13 +302,14 @@ WHERE (ServiceUser.UserCk IS NULL)) AS derivedtbl_1 ON Clean.ServiceFk = derived
                         var cleaning = (Cleaning)service;
                         var cleaningQuery = new SqlCommand("UPDATE Clean SET Size = @size, Remarks = @remarks WHERE CleanPk = @id");
                         cleaningQuery.Parameters.AddWithValue("@size", (int)cleaning.Size);
-                        cleaningQuery.Parameters.AddWithValue("@remarks", cleaning.Comments);
+                        cleaningQuery.Parameters.AddWithValue("@remarks", cleaning.Comments ?? "");
                         break;
                     }
             }
             var query = new SqlCommand("UPDATE Service SET StartDate = @startdate, EndDate = @enddate, TramFk = @tramfk WHERE ServicePk = @id");
             query.Parameters.AddWithValue("@startdate", service.StartDate);
-            query.Parameters.AddWithValue("@enddate", service.EndDate);
+            if (service.EndDate == DateTime.MinValue) query.Parameters.AddWithValue("@enddate", DBNull.Value);
+            else query.Parameters.AddWithValue("@enddate", service.EndDate);
             query.Parameters.AddWithValue("@tramfk", service.TramId);
             GetData(query);
             SetUsersToServices(service.AssignedUsers, service);
@@ -332,12 +335,13 @@ WHERE (Service.ServicePk = @id)");
 
         private static void SetUsersToServices(List<User> users, Service service)
         {
+            if (users == null) return;
             var data = GetData(new SqlCommand($"SELECT * FROM ServiceUser WHERE ServiceCk = {service.Id}"));
             foreach (DataRow dataRow in data.Rows)
             {
-                if (users.All(x => x.Id != (int) dataRow.ItemArray[0]))
+                if (users.All(x => x.Id != (int)dataRow.ItemArray[0]))
                 {
-                    GetData(new SqlCommand($"DELETE FROM ServiceUser WHERE ServiceCk = {service.Id} AND UserCk = {(int) dataRow.ItemArray[0]}"));
+                    GetData(new SqlCommand($"DELETE FROM ServiceUser WHERE ServiceCk = {service.Id} AND UserCk = {(int)dataRow.ItemArray[0]}"));
                 }
             }
